feat: Misc. cleanup, [gjstr] works with ISO-8859-1 standard. feat: Player [verify_id] is now a u16, passed to player creation.

+++ src/engine/src/io/server/model/if_opensub.rs
@@ -4,11 +4,11 @@ pub struct If_OpenSub {
     pub(crate) component_id: u32,
     pub(crate) interface_id: u32,
     pub(crate) flags: u32,
-    pub(crate) verify_id: u32,
+    pub(crate) verify_id: u16,
 }
 
 impl If_OpenSub {
-    pub fn new(window_id: u32, component_id: u32, interface_id: u32, flags: u32, verify_id: u32) -> If_OpenSub {
+    pub fn new(window_id: u32, component_id: u32, interface_id: u32, flags: u32, verify_id: u16) -> If_OpenSub {
         If_OpenSub {
             window_id,
             component_id,
